add now supports double data typr

+++ calc.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-int add(int x,int y){
+double add(double x,double y){
 	return x+y;
 }
 int sub(int x,int y){
